Refactor function getComputerChoice()
In the previous iteration, the process of randomly chosing one of the
three items consisted of establishing a random number from 1 through 99
and then dividing this range into three equally sized chunks, each
connected to one of the three items.

On Youtube I found a solution that is much more elegant:
- create an array containing the three items
- indexing into the array using a random integer within the range of 0
  through 2

I have adopted this approach.
The credits go to "Coding With Rob"
https://www.youtube.com/watch?v=HxSxmPDNXX4